fix(balance): refactoring wallet balance dropdown for correct wording and designs
