+++ packages/blockchain-wallet-v4-frontend/src/assets/locales/en.json
@@ -514,8 +514,8 @@
   "layouts.wallet.menuleft.navigation.borrow": "Borrow",
   "layouts.wallet.menuleft.navigation.earninterest": "Earn Interest",
   "layouts.wallet.menuleft.navigation.hardware": "Hardware",
-  "layouts.wallet.menutop.balance.walletbalance.showcrypto": "Show Crypto",
-  "layouts.wallet.menutop.balance.walletbalance.showfiat": "Show {currency}",
+  "layouts.wallet.menutop.balance.walletbalance.show.in.crypto": "Show in Crypto",
+  "layouts.wallet.menutop.balance.walletbalance.show.in.fiat": "Show in {currency}",
   "loader.message.gettingready": "Getting Ready...",
   "lockbox.service.messages.appalreadyinstalled": "App already installed",
   "lockbox.service.messages.btcrequired": "Unable to remove BTC app as it is required by others.",
