+++ packages/blockchain-wallet-v4-frontend/src/assets/locales/index.d.ts
@@ -616,8 +616,8 @@ type MessagesType = {
   'layouts.wallet.menuleft.navigation.swap': 'Swap'
   'layouts.wallet.menutop.balance.walletbalance.hardware': 'Hardware'
   'layouts.wallet.menutop.balance.walletbalance.nonspendable': 'Non-Spendable'
-  'layouts.wallet.menutop.balance.walletbalance.showcrypto': 'Show Crypto'
-  'layouts.wallet.menutop.balance.walletbalance.showfiat': 'Show {currency}'
+  'layouts.wallet.menutop.balance.walletbalance.show.in.crypto': 'Show in Crypto'
+  'layouts.wallet.menutop.balance.walletbalance.show.in.fiat': 'Show in {currency}'
   'layouts.wallet.menutop.balance.walletbalance.wallet': 'Wallet'
   'layouts.wallet.menutop.request': 'Request'
   'layouts.wallet.menutop.send': 'Send'
